requirements.txt
django==5.1.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
openpyxl>=3.1.2
celery==5.4.0
redis==5.2.1
pillow==11.0.0
django-allauth==65.4.0
drf-spectacular==0.28.0
rollbar==1.1.0
django-cacheops==7.1
django-grappelli==4.0.1
coverage==7.6.9
